(chore): move insight card to right folder
