+++ src/app/(protected)/dashboard/_components/trending-cards/insights-card.tsx
